Add files via upload
This is a real-time project that I did on Upwork as a freelancer. This dataset is related marketing domain . 
So here is the project description
Customer segmentation is the practice of dividing a customer base into groups of individuals that are similar in specific ways relevant to marketing, such as age, gender, interests and spending habits.

Companies employing customer segmentation operate under the fact that every customer is different and that their marketing efforts would be better served if they target specific, smaller groups with messages that those consumers would find relevant and lead them to buy something. Companies also hope to gain a deeper understanding of their customer's preferences and needs with the idea of discovering what each segment finds most valuable to more accurately tailor marketing materials toward that segment.

This weekend, we are back with another Janatahack, this time dealing in a problem statement on customer segmentation. .